Extract ethnicity for all patient
As mentioned in the comment, we now set the study population to all
patients because we only extracting ethnicity at one time point. If we
restrict this to our study population, cohorts extracted at another
time may not be included in this cohort. This led to categories of
both "Unknown" and "". Where we actually do have information on some of
the "" patients at other time points.

Opportunistic clean-up of extra spaces in project.yaml