Change to consistent use of index_date=start_date
For QOF studies we had previously used index_date=end_date to try to
align with the NHS finanical year. When running monthly we now use
last_day_of_month(index_date). Now that we are using patients.all() for
the ethnicity study definition, we can change this to be consistent with
OpenSAFELY documentation so it is less likely to cause confusion.