Fix learning disaibility and carehome labels
The measures framework requires that the denominator is a number and not
a string. Rather than do numerator/carehome with groupby=population, do
numerator/population with groupby=carehome. This means the measures file
will have a row for each group rather than just the subgroup, but we can
sort that later in plotting.