Extract LDA subgroup variables to input files
For learning disability and autism we extract both the binary
flag variable, and the varaible categorised with textual labels. We need
to use the binary value for the denominator of measures, but we can use
the labels as a "group_by" variable and for automatic labeling of plots
and tables. Instead of hiding the binary flag variables with nesting,
expose them to the input files.