Combine lda, register, dep003 study definitions
In the LDA study we now also want to look at the total population, not
just the learning disability and/or autism sub population. Because the
population is expanding from 300k to the whole registered population, it
is likely more efficient to extract the few extra QOF register and dep003
variables on the same cohort, rather than re-extracting all the
demographic variables for each person.

With a single study definition, update to a single output directory and
update the reporting with the new measures names