Split into three study definitions
The measures framework assumes that the numerator is a subset of the
denominator. When computing a numerator over a subgroup of the study
population, it does not filter for the subset of the numerator that is
relevant to the subgroup.

Split into three study definitions so we can be more precise with the
study population.