Check register for ongoing episode of depression
Business rules specify that those on the register have to have their
latest first or new episode of depression since the register date. If
there was a depression code entered prior to the register date with the
same code, and no resolved code between, the episode was still ongoing.

Compare the last match before the register date and the first match
after the register date.