Implement partially first/new diagnosis for dep003
Instead of using the latest depression diagnosis date, use the first
date in the last 15 months, excluding those who had ongoing depression

We are still potentially missing cases if the diagnosis changed during
the 15 months, so extract the next two diagnosis codes and count
changes. A complete analysis would require 10 codes.

We are also missing if there was a resolved code during the 15 months,
so count those.