Create subgroup variables and reformat measures
The measures framework is a ratio, not a rate (i.e. it does not ensure
that the numerator is a subset of the denominator, see
https://github.com/opensafely-core/cohort-extractor/issues/837). If we
want to look at subgroups in one study definition, one solution is to
create a new variable for the numerator which the the logical and of
the numerator and the denominator.

We add two loops, create_antidepressant_vars to simplify the creation of
variables identical except for the codelist and create_subgroups to
create the new numerator variables as described above.

Clean up code to remove unused dep003 variables and produce clearer
measure file names.