Move legend to right of plot
In order to automatically place legend somewhere it will not overlap
with data, place legend as a single column on the right side of the
plot. Use constrained_layout=True so subplots do not overlap.

Placement of matplotlib legend is based on of the left edge.
If we wanted to place the legend below the plot, we would need to know
how many legend columns would fit best, and we would need to know the
height of the legend given the number of columns selected.