Add groupby rather than using study def variables
For itsa analysis, we will want to use the remaining population i.e. not
ld or not a as a control group. Adding the subgroup variable as an
additional groupby variable means we no longer need to create extra
numerator variables for autism and antidepressants or learning
disability an antidepressants (measures framework is a ratio not a
rate).