Update to allow for multiple groupby
Panel_plots and join_and_round were using the hardcoded columns of group
and category. Panel plots now generates indexed groups and categories
i.e. group_0 and category_0.

For a given subset of measures, if the category is all the
same and the group is a boolean, filter where the rows are True.
Also create new category and group columns from the last category and
group seen.

This is useful when we have a few subgroups and then a set of breakdowns
i.e. groupby = ["autism"], ["imd"], ["autism", "imd"], and the subgroup
is listed before the breakdown.