Adjust measures for new
Using 2 years to check for a previous depression diagnosis does not make
sense, as the average number of depression codes per person is 2-3
(unlike prescribing which could happen every 3 months). It would be
difficullt to know who was on the depression register last month without
re-extracting each of the register variables again offset by a month.
Otherwise it would need a separate analysis script.

Instead focus on the diagnosis of prescribing versus new prescribing.

Add measures for demographic breakdowns (including diagnosis) of new
antidepressant prescribing. They may have lots of redaction, but could
exclude from future plots if that is the case.