Split prescription subtype in one measure
Instead of having separate measures for the different prescription types
(which are hard to analyse), create one breakdown

There are around 0.5% to 1% of people with multiple prescription types
in one month. Separate those out as another group.

The measures framework does not have a way to do a groupby that does not
split the denominator, so create a measure where the numerator and
denominator are the same to just get a count. Update join_and_round to
recreate the denominator based on the sum of the numerator.