Add computation of geometric mean
To have a summary effect size compute the geometric mean as described
in "Effect size quantification for interrupted time series analysis:
implementation in R and analysis for Covid-19 research"

In order to display for the demographic breakdown models, move towards
using a single model for each subgroup, rather than an interaction.

In the process of updating the tables and figures.