Choose different month for table1
We choose a representitive month towards the end of the study period
instead of the first month of the study period (January), which also
might have more prescriptions (post-holiday month).

Also allow option to separately choose whether to display denominator
and rate. Now we just want rate, not denominator (to save space).